fix subscribe for new users
Use new add_subscriber logic.
This means correctly determining if a user is already subscribed and handling any state correctly.

Note that this code is still not perfect, as it has some custom functionality for subscribing as a non-user, but it is also just named "add_subscriber".
This needs a small refactor to be more clear about the distinction.