don't allow unsubscribe if pending
Mailchimp does not allow to "delete" (archive) a user that is pending. This means it is not possible for the user to remove themselves via the websites. Instead they will need to contact NAP to do this manually.
Disable the checkbox to make it less confusing why unchecking it has no effect.