fix resubscribing in short amount of time
The best way to resubscribe would be to set the user to pending, but mailchimp will not resend a confirmation mail if it did it too soon already (could be 24hour time window).
This can be a problem, as if a user unsubscribes and resubscribes in this time window, they will never receive the email and will be stuck in the pending status.

Instead, as the user agreed with the checkbox, resubscribe directly.